requirements.txt
discord.py==2.2.2
emoji==2.2.0
openai==0.27.0
